# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/subscribe":{"post":{"summary":"Subscribe","operationId":"subscribe_subscribe_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscribeItem"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/unsubscribe":{"post":{"summary":"Unsubscribe","operationId":"unsubscribe_unsubscribe_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnsubscribeItem"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/getSubscription":{"get":{"summary":"Get Subscription","operationId":"get_subscription_getSubscription_get","parameters":[{"name":"camera_url","in":"query","required":true,"schema":{"type":"string","title":"Camera Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/checkCameraConnection":{"get":{"summary":"Connect Camera","operationId":"connect_camera_checkCameraConnection_get","parameters":[{"name":"camera_url","in":"query","required":true,"schema":{"type":"string","title":"Camera Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/updateEnv":{"put":{"summary":"Update Env","operationId":"update_env_updateEnv_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnvItem"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EnvItem":{"properties":{"min_contour_area":{"type":"integer","title":"Min Contour Area"},"cooldown_seconds":{"type":"integer","title":"Cooldown Seconds"}},"type":"object","required":["min_contour_area","cooldown_seconds"],"title":"EnvItem"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SubscribeItem":{"properties":{"movement_callback_url":{"type":"string","title":"Movement Callback Url"},"crash_callback_url":{"type":"string","title":"Crash Callback Url"},"camera_url":{"type":"string","title":"Camera Url"}},"type":"object","required":["movement_callback_url","crash_callback_url","camera_url"],"title":"SubscribeItem"},"UnsubscribeItem":{"properties":{"camera_url":{"type":"string","title":"Camera Url"}},"type":"object","required":["camera_url"],"title":"UnsubscribeItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}